feat(identifyCustomDelimiter): Add custom delimiter identification and update delimiter list
Added functionality to identify a custom delimiter in the input string. If a custom
delimiter is detected, it is added to the default delimiter list (`,`, `:`) for processing.
This allows input strings with custom delimiters to be correctly split and processed.

BREAKING CHANGE: Input strings that specify a custom delimiter using the `//` and `\n`
notation will now have that delimiter added to the list of delimiters used for splitting the string.